--- a/bestla/bestla/bestla_parallel.h
+++ b/bestla/bestla/bestla_parallel.h
@@ -137,23 +137,25 @@ class StdThreading : public IThreading {
   }
 
   inline void sync(int tidx, int idx = 0) override {
-    flag[idx].fetch_sub(1);
-    if (cr->mHybrid) {
-      Timer_T tm;
-      tm.start();
-      while (true) {
-        if (flag[idx].load() == 0)
-          break;
-        else
-          _mm_pause();
-      }
-      thread_time[tidx] -= int(tm.stop());
-    } else {
-      while (true) {
-        if (flag[idx].load() == 0)
-          break;
-        else
-          _mm_pause();
+    if (mThreadNum > 1) {
+      flag[idx].fetch_sub(1);
+      if (cr->mHybrid) {
+        Timer_T tm;
+        tm.start();
+        while (true) {
+          if (flag[idx].load() == 0)
+            break;
+          else
+            _mm_pause();
+        }
+        thread_time[tidx] -= int(tm.stop());
+      } else {
+        while (true) {
+          if (flag[idx].load() == 0)
+            break;
+          else
+            _mm_pause();
+        }
       }
     }
   }