--- a/bestla/bestla/bestla_utils.h
+++ b/bestla/bestla/bestla_utils.h
@@ -90,9 +90,7 @@
 #define CompileAMXINT8() (CompileAMX())
 #endif
 
-#if CompileBF16() || CompileFP16()
 #include <immintrin.h>
-#endif
 
 namespace bestla {
 namespace utils {