## Type of Change

Get higher quantization accuracy of BesTLA's quantization packweight API.

- [x] Introduce auto-fullrange for NBits quantization
- [x] Add int3 rounding conversion
- [x] Optimize int4 decompression on client CPUs
- [x] Remove S4_Fullrange, as it's already covered by auto-fullrange
- [x] root cause Float4 performance issue on hybrid CPUs, 20%+ speedup