@luoyu-intel 2024-03-13
Text generation comparison ( weight_dtype=int3, group_size=128)

This PR:
 Once upon a time, there existed a little girl, who liked to have adventures. She wanted to go to places and meet new people, and have fun. And every day that same little girl would put on her best dress and pear
model_print_timings:        load time =    73.59 ms
model_print_timings:      sample time =     8.07 ms /    16 runs   (    0.50 ms per token)
model_print_timings: prompt eval time =    73.56 ms /    34 tokens (    2.16 ms per token)
model_print_timings:        eval time =   285.11 ms /    15 runs   (   19.01 ms per token)
model_print_timings:       total time =   370.79 ms
========== eval time log of each prediction ==========
prediction   0, time: 73.56ms
prediction   1, time: 19.65ms
prediction   2, time: 19.07ms
prediction   3, time: 18.98ms
prediction   4, time: 19.01ms
prediction   5, time: 19.08ms

Main:
 Once upon a time, there existed a little girl, who liked to have adventures. She wanted to go to places and meet new people, and have fun. Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија Хронологија
model_print_timings:        load time =    73.01 ms
model_print_timings:      sample time =     7.62 ms /    16 runs   (    0.48 ms per token)
model_print_timings: prompt eval time =    72.98 ms /    34 tokens (    2.15 ms per token)
model_print_timings:        eval time =   289.00 ms /    15 runs   (   19.27 ms per token)
model_print_timings:       total time =   373.55 ms
========== eval time log of each prediction ==========
prediction   0, time: 72.98ms
prediction   1, time: 19.92ms
prediction   2, time: 19.21ms
prediction   3, time: 19.20ms
prediction   4, time: 19.23ms
prediction   5, time: 19.45ms