@luoyu-intel 2024-03-13
@kevinintel @hshen14  INT3 RTN quantization can generate reasonable texts now. I added 'int3' to the quantization weight_dtype options in this PR.